feat: add searchable field to jwt.startegy.base
